We don't need only two methods for the programs controller to be in a separate file.
