+++ app/controllers/programs_controller.rb
@@ -5,7 +5,6 @@
 class ProgramsController < ApplicationController
 
   include Concern::Controller::GetPopularArticles
-  include Concern::Controller::ShowEpisodes
 
   layout 'new/single', only: [:segment]
 
@@ -177,6 +176,21 @@ class ProgramsController < ApplicationController
 
 
   private
+
+  def render_kpcc_episode
+    @content = @episode.published_content.map(&:to_article)
+    @featured_programs = KpccProgram.where.not(id: @program.id, is_featured: false).first(4)
+    if @program.is_segmented?
+      @episodes = @program.episodes.published.order("air_date").first(4)
+      render 'programs/kpcc/episode', layout: 'new/ronin' and return
+    else
+      render 'programs/kpcc/old/episode_standalone'
+    end
+  end
+  def render_external_episode
+    @segments = @episode.segments
+    render 'programs/external/episode', layout: 'application' and return
+  end
 
   def get_program
     @program = KpccProgram.find_by_slug(params[:show]) ||
